requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
cssselect2==0.5.0
dj-database-url==1.0.0
Django==3.2.15
django-bootstrap-v5==1.0.11
django-bootstrap4==22.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-embed-video==1.4.1
django-filter==21.1
django-heroku==0.3.1
django-multiselectfield==0.1.12
djangorestframework==3.13.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==2.1.3
lxml==4.8.0
oscrypto==1.3.0
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
pyOpenSSL==22.0.0
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
razorpay==1.3.0
reportlab==3.6.9
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.3.1
stripe==2.67.0
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.64.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.8
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.7
zipp==3.8.0
